- refator : 주문, 상품 재고 entity 수정

+++ src/main/java/kr/hhplus/be/server/domain/product/entity/ProductSalesSummary.java
@@ -32,7 +32,7 @@ public class ProductSalesSummary {
         this.orderedAt = orderedAt;
     }
 
-    public void increaseQty(long qty) {
+    public void increaseQty(int qty) {
         this.totalQty += qty;
     }
 }
